Move helper functions to utility library & gNMI port interface tests
