+++ tests/qos/cpu_qos_test.cc
@@ -67,10 +67,10 @@
 #include "p4_pdpi/pd.h"
 #include "p4_pdpi/string_encodings/decimal_string.h"
 #include "proto/gnmi/gnmi.pb.h"
-#include "sai_p4/instantiations/google/sai_p4info.h"
 #include "sai_p4/instantiations/google/sai_pd.pb.h"
 #include "tests/forwarding/util.h"
 #include "tests/qos/gnmi_parsers.h"
+#include "tests/qos/qos_test_util.h"
 #include "thinkit/control_device.h"
 #include "thinkit/generic_testbed.h"
 #include "thinkit/mirror_testbed.h"
@@ -237,7 +237,7 @@ absl::StatusOr<p4::v1::TableEntry> SetUpPuntToCPUWithRateLimit(
             src_ip { value: "$1" mask: "255.255.255.255" }
             dst_ip { value: "$2" mask: "255.255.255.255" }
           }
-          action { acl_experimental_trap { qos_queue: "$3" } }
+          action { acl_trap { qos_queue: "$3" } }
           priority: 1
           meter_config { bytes_per_second: $4 burst_bytes: $5 }
         }
@@ -250,100 +250,6 @@ absl::StatusOr<p4::v1::TableEntry> SetUpPuntToCPUWithRateLimit(
                    pdpi::PartialPdTableEntryToPiTableEntry(ir_p4info, acl_entry));
   RETURN_IF_ERROR(pdpi::InstallPiTableEntry(&p4_session, pi_acl_entry));
   return pi_acl_entry;
-}
-
-// These are the counters we track in these tests.
-struct QueueCounters {
-  int64_t num_packets_transmitted = 0;
-  int64_t num_packet_dropped = 0;
-};
-
-std::ostream &operator<<(std::ostream &os, const QueueCounters &counters) {
-  return os << absl::StreamFormat(
-             "QueueCounters{"
-             ".num_packets_transmitted = %d, "
-             ".num_packets_dropped = %d"
-             "}",
-             counters.num_packets_transmitted, counters.num_packet_dropped);
-}
-
-// TODO: Move this to a helper library.
-absl::StatusOr<QueueCounters> GetGnmiQueueCounters(
-    absl::string_view port, absl::string_view queue,
-    gnmi::gNMI::StubInterface &gnmi_stub) {
-  QueueCounters counters;
-  const std::string openconfig_transmit_count_state_path = absl::Substitute(
-      "qos/interfaces/interface[interface-id=$0]"
-      "/output/queues/queue[name=$1]/state/transmit-pkts",
-      port, queue);
-
-  ASSIGN_OR_RETURN(
-      std::string transmit_counter_response,
-      GetGnmiStatePathInfo(&gnmi_stub, openconfig_transmit_count_state_path,
-                           "openconfig-qos:transmit-pkts"));
-
-  if (!absl::SimpleAtoi(StripQuotes(transmit_counter_response),
-                        &counters.num_packets_transmitted)) {
-    return absl::InternalError(absl::StrCat("Unable to parse counter from ",
-                                            transmit_counter_response));
-  }
-
-  const std::string openconfig_drop_count_state_path = absl::Substitute(
-      "qos/interfaces/interface[interface-id=$0]"
-      "/output/queues/queue[name=$1]/state/dropped-pkts",
-      port, queue);
-
-  ASSIGN_OR_RETURN(
-      std::string drop_counter_response,
-      GetGnmiStatePathInfo(&gnmi_stub, openconfig_drop_count_state_path,
-                           "openconfig-qos:dropped-pkts"));
-
-  if (!absl::SimpleAtoi(StripQuotes(drop_counter_response),
-                        &counters.num_packet_dropped)) {
-    return absl::InternalError(
-        absl::StrCat("Unable to parse counter from ", drop_counter_response));
-  }
-
-  return counters;
-}
-
-// Returns the total number of packets enqueued for the queue with the given
-// `QueueCounters`.
-int64_t CumulativeNumPacketsEnqueued(const QueueCounters &counters) {
-  return counters.num_packet_dropped + counters.num_packets_transmitted;
-}
-
-absl::Status SetPortSpeed(const std::string &port_speed,
-                          const std::string &iface,
-                          gnmi::gNMI::StubInterface &gnmi_stub) {
-  std::string ops_config_path = absl::StrCat(
-      "interfaces/interface[name=", iface, "]/ethernet/config/port-speed");
-  std::string ops_val =
-      absl::StrCat("{\"openconfig-if-ethernet:port-speed\":", port_speed, "}");
-  RETURN_IF_ERROR(pins_test::SetGnmiConfigPath(&gnmi_stub, ops_config_path,
-                                               GnmiSetType::kUpdate, ops_val));
-  return absl::OkStatus();
-}
-
-absl::Status SetPortMtu(int port_mtu, const std::string &interface_name,
-                        gnmi::gNMI::StubInterface &gnmi_stub) {
-  std::string config_path = absl::StrCat(
-      "interfaces/interface[name=", interface_name, "]/config/mtu");
-  std::string value = absl::StrCat("{\"config:mtu\":", port_mtu, "}");
-  RETURN_IF_ERROR(pins_test::SetGnmiConfigPath(&gnmi_stub, config_path,
-                                               GnmiSetType::kUpdate, value));
-  return absl::OkStatus();
-}
-
-absl::StatusOr<bool> CheckLinkUp(const std::string &iface,
-                                 gnmi::gNMI::StubInterface &gnmi_stub) {
-  std::string oper_status_state_path =
-      absl::StrCat("interfaces/interface[name=", iface, "]/state/oper-status");
-  std::string parse_str = "openconfig-interfaces:oper-status";
-  ASSIGN_OR_RETURN(
-      std::string ops_response,
-      GetGnmiStatePathInfo(&gnmi_stub, oper_status_state_path, parse_str));
-  return ops_response == "\"UP\"";
 }
 
 absl::StatusOr<packetlib::Packet> MakeIpv4PacketWithDscp(
@@ -408,27 +314,6 @@ absl::StatusOr<packetlib::Packet> MakeIpv6PacketWithDscp(
   RETURN_IF_ERROR(packetlib::PadPacketToMinimumSize(packet).status());
   RETURN_IF_ERROR(packetlib::UpdateAllComputedFields(packet).status());
   return packet;
-}
-
-absl::StatusOr<absl::flat_hash_map<int, std::string>>
-ParseIpv4DscpToQueueMapping(absl::string_view gnmi_config) {
-  // TODO: Actually parse config -- hard-coded for now.
-  absl::flat_hash_map<int, std::string> queue_by_dscp;
-  for (int dscp = 0; dscp < 64; ++dscp) queue_by_dscp[dscp] = "BE1";
-  for (int dscp = 8; dscp <= 11; ++dscp) queue_by_dscp[dscp] = "AF1";
-  queue_by_dscp[13] = "LLQ1";
-  for (int dscp = 16; dscp <= 19; ++dscp) queue_by_dscp[dscp] = "AF2";
-  queue_by_dscp[21] = "LLQ2";
-  for (int dscp = 24; dscp <= 27; ++dscp) queue_by_dscp[dscp] = "AF3";
-  for (int dscp = 32; dscp <= 35; ++dscp) queue_by_dscp[dscp] = "AF4";
-  for (int dscp = 48; dscp <= 59; ++dscp) queue_by_dscp[dscp] = "NC1";
-  return queue_by_dscp;
-}
-
-absl::StatusOr<absl::flat_hash_map<int, std::string>>
-ParseIpv6DscpToQueueMapping(absl::string_view gnmi_config) {
-  // TODO: Actually parse config -- hard-coded for now.
-  return ParseIpv4DscpToQueueMapping(gnmi_config);
 }
 
 // Represents a link connecting the switch under test (SUT) to a control device.
